pyproject.toml
[project]
name = "my_google_search"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp",
    "googlesearch-python"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["my_google_search"]
